requirements.txt
ads
arxiv
pylatexenc
bibtexparser
requests
anytree
flask
PyYAML
